add sphinx-gallery uber h3 example
