+++ docs/src/conf.py
@@ -203,8 +203,7 @@ else:
 sphinx_gallery_conf = {
     "filename_pattern": "/.*",
     "ignore_pattern": (
-        "(__init__)|(clouds)|(earthquakes)|(from_1d)|(from_2d)"
-        "|(from_unstructured)|(uber)"
+        "(__init__)|(clouds)|(earthquakes)|(from_1d)|(from_2d)" "|(from_unstructured)"
     ),
     "examples_dirs": "../../src/geovista/examples",
     "gallery_dirs": "generated/gallery",
@@ -215,4 +214,7 @@ sphinx_gallery_conf = {
     "download_all_examples": False,
     "remove_config_comments": True,
     "within_subsection_order": ExampleTitleSortKey,
+    "reference_url": {
+        "geovista": None,
+    },
 }
