+++ src/geovista/examples/README.rst
@@ -1,6 +1,7 @@
 .. _examples_gallery:
 
+****************
 Examples Gallery
-================
+****************
 
 An interactive gallery of the ``geovista.examples``.
