+++ src/geovista/examples/point_cloud/README.rst
@@ -1,2 +1,2 @@
 Point Cloud
------------
+===========
