+++ src/geovista/examples/spatial_index/README.rst
@@ -1,0 +1,2 @@
+Spatial Index
+=============
