+++ src/geovista/registry.txt
@@ -47,8 +47,6 @@ tests/images/from_2d__synthetic_face_m1n1.png 9ef84bc7a41d237063f819b56a122b943e
 tests/images/from_2d__synthetic_face_m1n1_robin.png c23335fd376d81bf47f9d9f70059fe0f03fd52b243426ea693e65ef9195340fd
 tests/images/from_2d__synthetic_node_m1n1_moll.png d93e9907bd306b95c2e67b5ed99466f08edb3cdd35de85bb4026076422f9b5e3
 tests/images/from_2d__synthetic_node_m1n1.png 9d00923e48dad39ba03d903b18a5151ffb085b98e8eba4ad0601ad8630aef5b8
-tests/images/point_cloud__from_points__orca_cloud_eqc.png 6421665bc4bff1128b2876b7b83978b3c6e894be607b39764e136f013b795ddc
-tests/images/point_cloud__from_points__orca_cloud.png a5fa2a43eb6f8b1ab2adf6f6e757830b07305a87a3bb9abf64c48d2d2d8a066b
 tests/images/from_unstructured__fesom_fouc.png 8a8de06cd8f18e9758c05baf3bad77b570bca883cb7f5f2ef4f3c3164e7cfe9d
 tests/images/from_unstructured__fesom.png 3a5463d807a8951072a33e0b8f9f83eedfea8ab9b41cd1a13ae4689478013e39
 tests/images/from_unstructured__fvcom.png d9a51935c932fafa9f45989e500e8da0273bdd0f8f3da9d1f581107aca9108cd
@@ -66,6 +64,8 @@ tests/images/from_unstructured__smc.png 976987bcfdc25610b085efa2c538e2ff9125f0e5
 tests/images/from_unstructured__smc_sinu.png c4e05cf1d4a8bea4da32aecb21a053a929198387c7722049bb8356e6b353da85
 tests/images/from_unstructured__tri_hammer.png 722c9e22881254547d415c2acc9a9b2fd113d1171152d770ca267bbf71420ec5
 tests/images/from_unstructured__tri.png 351847600795a0c1569e71414786311091d1f520915452c3def67b7b03a9244b
-tests/images/uber_h3.png 08cb9cbb5f3ae41d1955efe7cdcfac999b2c6e67072f57871dde85ab49205289
+tests/images/point_cloud__from_points__orca_cloud_eqc.png 6421665bc4bff1128b2876b7b83978b3c6e894be607b39764e136f013b795ddc
+tests/images/point_cloud__from_points__orca_cloud.png a5fa2a43eb6f8b1ab2adf6f6e757830b07305a87a3bb9abf64c48d2d2d8a066b
+tests/images/spatial_index__uber_h3.png 08cb9cbb5f3ae41d1955efe7cdcfac999b2c6e67072f57871dde85ab49205289
 tests/images/vector_data__vectors.png fdfea84491039fa742b73e610b834272d73eac7a70ab6b8c6e9b51019c0e1a95
 tos_Omon_AWI-ESM-1-1-LR_historical_r1i1p1f1_gn_185001-185012.nc 28a8f7b8f3d8edded39a7b9c93afe54aeefbf271086927c7f684207327acc882
