+++ tests/plotting/test_examples.py
@@ -51,9 +51,9 @@ if not cache_dir.exists():
 
 # individual GHA CI example test case exceptions to the default image tolerances
 thresholds = {
-    "from_points__orca_cloud": {"warning_value": 202.0},
-    "from_points__orca_cloud_eqc": {"warning_value": 250.0},
-    "uber_h3": {"warning_value": 395.0},
+    "point_cloud.from_points__orca_cloud": {"warning_value": 202.0},
+    "point_cloud.from_points__orca_cloud_eqc": {"warning_value": 250.0},
+    "spatial_index.uber_h3": {"warning_value": 395.0},
 }
